+++ src/main/workers/addToRecentProjects/index.ts
@@ -27,7 +27,9 @@ ipcMain.on('addToRecentProjects', async (e: IpcMainEvent, { projectName, filePat
     path: filePath,
   });
 
-  await fsPromises.unlink(recentProjectsFilePath);
+  if (fs.existsSync(recentProjectsFilePath)) {
+    await fsPromises.unlink(recentProjectsFilePath);
+  }
   await fsPromises.writeFile(recentProjectsFilePath, JSON.stringify(recentProjects), 'utf8');
   await fsPromises.unlink(pathToLockFile);
